www/ui: Extract exported functions to separate file to fix fast reloads
